renamed package to hold entities
